# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Expoza API",
        "description": "Expoza API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "List all orders for the authenticated user",
                "description": "List all orders for the authenticated user",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ongoing",
                                "past"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Show a single order's details",
                "description": "Show a single order's details",
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}/refund": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Request a refund for an order",
                "description": "Request a refund for an order",
                "operationId": "09a90c074a0f3568ad9364d85af17f75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Refund request submitted successfully."
                    },
                    "403": {
                        "description": "This order is not eligible for a refund."
                    },
                    "409": {
                        "description": "A refund request has already been submitted for this order."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}/invoice": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get invoice data for an order",
                "description": "Get invoice data for an order",
                "operationId": "c12caa8a2a52b417459c2a84156eb444",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/past-orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get past orders",
                "operationId": "4c6a5bfae5fc8a00cb6461163d7a2ece",
                "responses": {
                    "200": {
                        "description": "List of past orders"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/place": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Place new order",
                "operationId": "4b93bb22d9f807e9add55efec55832a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address_id",
                                    "payment_method"
                                ],
                                "properties": {
                                    "address_id": {
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash_on_delivery",
                                            "credit_card",
                                            "paypal"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order placed successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/cancel/{id}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Cancel order",
                "operationId": "c871e33595e6c2ba825ea7f0d0167de2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "422": {
                        "description": "Order cannot be cancelled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}